Refactor Kostal Plenticore tests to consolidate mock settings and improve fixture usage
